Refactor model specs to use build not create
By using Factory girl in the test, we are essentially testing the behavior of Teacher, User, and the Factory for Teacher.

The purpose of this test is to test ONLY that a course delegates the last_name method to teacher.

By creating a double that simply responds to `#last_name`, we can ensure that we are testing course in isolation.

We were also hitting the database by using create and not build. This is unnecessary for this particular test.