Move course delete button to `edit` page.
New courses don't have an `id` so they can't be deleted. The Rails route
helpers, when building a `delete` link, requires an `id`, and fails when
the model it's using doesn't have one.

There were, additionally, a few locales that were mistakenly removed during
a previous rebase. The only one that broke things was `helpers.delete_model`.
Added that back in to make things green(er).